move commands to commands module
